+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- actors are now only considered in combat, if they are part of a started combat
+  in the active scene or part of an unlinked, started combat
+  ([#210](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/210))
+
 ### Fixed
 
 - some caliber types not localizing correctly